Onboarding analyze API
