+++ .env.test
@@ -1,4 +1,5 @@
-THERESA_API_URL=https://theresa.nexusgraph.com/v1/data/
+ANALYZE_API_RESOURCE=https://TODO-add-mock-server
+GRAPH_API_RESOURCE=http://localhost:8080/v1/data
 
 DEV_USER_ID=10000
 
@@ -8,6 +9,6 @@ LOGTO_SIGN_OUT_REDIRECT_URL=http://localhost:3000/login
 LOGTO_APP_ID=ypon89z8rtrjdg5ta669l
 TEST_USER_EMAIL=testUser@test.com
 TEST_USER_PASSWORD=ojeHDMRQ
-GRAPH_API_RESOURCE=http://localhost:8080/v1/data
+
 SKIP_SIGN_IN=true
 NODE_ENV=development
