+++ cypress/fixtures/single-rdf-tuple-from-analysis.json
