+++ docs/i18n/zh-cn/docusaurus-plugin-content-docs/current/development/5-environment.md
@@ -56,7 +56,7 @@ npm install --global yarn
 创建一个包含 Nexus Graph 运行时需要的所有变量的 [`.env` 文件][`.env` file]。以下变量
 需要定义:
 
-- **THERESA_API_URL** 请求 [Theresa API](https://theresa-api.com) 的URL，用于将将自然语言文本转化为知识图谱
+- **ANALYZE_API_RESOURCE** 请求 [Theresa API](https://theresa-api.com) 的URL，用于将将自然语言文本转化为知识图谱
 - **LOGTO_ENDPOINT_URL**
   - [Logto](https://docs.logto.io/) 提供了一个全面的身份验证的解决方案，包括前端和后端，以及预构建的基础配置和企业级解决方案。
   - 在 Nexus Graph 中，我们使用Logto来验证用户是否已登录并自动生成用户登录页面
@@ -80,7 +80,7 @@ npm install --global yarn
 [`.env.test` 文件][`.env.test` file] 中提供了一个 [`.env` 文件] 的例子：
 
 ```bash
-THERESA_API_URL=http://localhost:5000/
+ANALYZE_API_RESOURCE=http://localhost:5000/
 LOGTO_ENDPOINT_URL=https://u4v5ne.logto.app/
 LOGTO_SIGN_IN_CALLBACK_URL=http://localhost:8080/login
 LOGTO_APP_ID=ypon89z8rtrjdg5ta669l
