+++ package.json
@@ -42,8 +42,6 @@
     "i18next": "^23.7.6",
     "i18next-browser-languagedetector": "^7.2.0",
     "immer": "^10.0.3",
-    "inversify": "^6.0.1",
-    "inversify-inject-decorators": "^3.1.0",
     "lodash-es": "^4.17.21",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -52,7 +50,6 @@
     "react-router-dom": "^6.8.1",
     "react-spring": "^9.7.3",
     "redux": "^3.7.2",
-    "reflect-metadata": "^0.1.13",
     "styled-components": "^5.3.9"
   },
   "devDependencies": {
