--- packages/nexusgraph-app/types.ts
@@ -1,7 +1,0 @@
-// Copyright 2023 Paion Data. All rights reserved.
-
-const TYPES = {
-  NLPClient: Symbol("NLPClient"),
-};
-
-export default TYPES;
