+++ packages/nexusgraph-nlp/index.ts
@@ -1,3 +1,2 @@
 // Copyright 2023 Paion Data. All rights reserved.
-export type { NLPClient } from "./src/NLPClient";
-export { TheresaClient } from "./src/TheresaClient";
+export { NLPClient } from "./src/NLPClient";
